add rest client generate; use transaction by sprint endpoint

+++ server/src/sprints/sprints.service.ts
@@ -3,13 +3,15 @@ import { CreateSprintDto } from './dto/create-sprint.dto';
 import { UpdateSprintDto } from './dto/update-sprint.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { EnvelopesService } from 'src/envelopes/envelopes.service';
+import { Prisma } from '@prisma/client';
+import { SprintDetailedResponseDto } from './dto/sprint.response.dto';
 
 @Injectable()
 export class SprintsService {
   constructor(
     private prisma: PrismaService,
     private envelopesService: EnvelopesService,
-  ) {}
+  ) { }
   async create(createSprintDto: CreateSprintDto, userId: number) {
     const { envelopes, ...createSprintData } = createSprintDto;
     const { id } = await this.prisma.sprint.create({
@@ -27,49 +29,25 @@ export class SprintsService {
     return Promise.all(createEnvelopesPromises).then(() =>
       this.prisma.sprint.findUnique({
         where: { id },
-        include: { envelopes: true },
       }),
     );
+
   }
 
   async findAll(userId: number) {
-    const sprints = await this.prisma.sprint.findMany({
+    return this.prisma.sprint.findMany({
       where: { userId },
-      include: {
-        transactions: {
-          select: { amount: true },
-        },
-        envelopes: {
-          select: { amount: true },
-        },
-      },
-    });
-    return sprints.map((s) => {
-      const totalSpendings = s.transactions.reduce(
-        (prev, curr) => curr.amount + prev,
-        0,
-      );
-      const totalPlain = s.envelopes.reduce(
-        (prev, curr) => curr.amount + prev,
-        0,
-      );
-      const { envelopes, transactions, ...rest } = s;
-      return {
-        ...rest,
-        totalSpendings,
-        totalPlain,
-      };
     });
   }
 
   async findOne(id: number, userId: number) {
-    await this.checkIsUsersSprintOrThrow(id, userId);
-    const findedSprint = await this.prisma.sprint.findUnique({
-      where: { id },
+    const foundedSprint = await this.prisma.sprint.findUnique({
+      where: { id, userId },
       include: {
         envelopes: {
           include: {
             transactions: true,
+            category: true
           },
           orderBy: {
             category: {
@@ -80,14 +58,16 @@ export class SprintsService {
         transactions: true,
       },
     });
-    const { transactions: _, ...result } = findedSprint;
-    const totalSpendings = findedSprint.transactions.reduce(
+    const { transactions: _, ...result } = foundedSprint;
+    const totalSpendings = foundedSprint.transactions.reduce(
       (prev, curr) => curr.amount + prev,
       0,
     );
     return {
       ...result,
-      currentBalance: findedSprint.startSum - totalSpendings,
+      totalSpendings,
+      totalPlain: foundedSprint.startSum,
+      currentBalance: foundedSprint.startSum - totalSpendings,
     };
   }
 
@@ -110,22 +90,18 @@ export class SprintsService {
   }
 
   async update(id: number, updateSprintDto: UpdateSprintDto, userId: number) {
-    await this.checkIsUsersSprintOrThrow(id, userId);
-    return this.prisma.sprint.update({ where: { id }, data: updateSprintDto });
+    try {
+      return this.prisma.sprint.update({ where: { id, userId }, data: updateSprintDto })
+    } catch (err) {
+      throw new NotFoundException('Sprint not found');
+    };
   }
 
   async remove(id: number, userId: number) {
-    await this.checkIsUsersSprintOrThrow(id, userId);
-    return this.prisma.sprint.delete({ where: { id } });
-  }
-
-  async checkIsUsersSprintOrThrow(sprintId: number, userId: number) {
-    const sprint = await this.prisma.sprint.findUnique({
-      where: { id: sprintId },
-    });
-    if (!sprint || sprint.userId !== userId) {
+    try {
+      return this.prisma.sprint.delete({ where: { id, userId } });
+    } catch (err) {
       throw new NotFoundException('Sprint not found');
     }
-    return sprint;
   }
 }
